% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSampleSize.R
\name{estimateSampleSize}
\alias{estimateSampleSize}
\title{Sample size estimate}
\usage{
estimateSampleSize(effect = 0.1, sigLevel = 0.05, powerLevel = 0.8)
}
\arguments{
\item{effect}{Numeric variable indicating effect size to discern}

\item{sigLevel}{Numeric variable indicating the level of significance}

\item{powerLevel}{Numeric indicator of the power level}
}
\value{
power Data frame indicating minimum sample sizes for various statistical tests
}
\description{
\code{estimateSampleSize(effect, sigLevel, powerLevel)}
}
\details{
Function to compute minimum sample size required for
Variable and partial response rates may affect sample size. To ensure that
readers could rely upon non-significant results, this analyst conducted several
power analyses using the pwr package (Ekstrom et al. 2017). Each test
indicated the minimum sample size required to detect a small effect
of 0.1, with the p<.05 probability of a type I error of p < .05
and power of 0.8
}
\author{
John James, \email{jjames@datasciencestudio.org}
}
