% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getMode}
\alias{getMode}
\title{getMode}
\usage{
getMode(v)
}
\arguments{
\item{v}{- vector of numbers}
}
\value{
mode - Numeric, the most frequently occurring value
}
\description{
\code{getMode(v)} returns the mode for a vector v
}
\details{
This function takes as its parameters, a vector of numeric values and
computes the mode, and returns it to the calling environment.
}
\author{
John James, \email{jjames@datasciencestudio.org}
}
