% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRq1.R
\name{analyzeRq1}
\alias{analyzeRq1}
\title{analyzeRq1}
\usage{
analyzeRq1(brfss)
}
\arguments{
\item{brfss}{Data frame of preprocessed 2013 Behavioral Risk Factor
Surveillance System (BRFSS) telephone survey.}
}
\value{
analysis List containing frequency and proportion data, plots,
    and statistical tests.
}
\description{
\code{analyzeRq1(brfss)}{Performs analysis of research question 1}
}
\details{
Function that performs the analysis for research question #1:  Is there an
association between education and income and the prevalence of depression?
}
\author{
John James, \email{jjames@dataference.com}
}
