% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRq2.R
\name{analyzeRq2}
\alias{analyzeRq2}
\title{analyzeRq2}
\usage{
analyzeRq2(brfss)
}
\arguments{
\item{brfss}{Data frame of preprocessed 2013 Behavioral Risk Factor
Surveillance System (BRFSS) telephone survey.}
}
\value{
analysis List containing frequency and proportion data, plots,
    and statistical tests.
}
\description{
\code{analyzeRq2(brfss)}{Performs analyses for research question #2}
}
\details{
Function that performs the analysis for research question #2:  Is there a
relationship between depression and chronic illness?
}
\author{
John James, \email{jjames@dataference.com}
}
