% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRq3.R
\name{analyzeRq3}
\alias{analyzeRq3}
\title{analyzeRq3}
\usage{
analyzeRq3(brfss)
}
\arguments{
\item{brfss}{Data frame of preprocessed 2013 Behavioral Risk Factor
Surveillance System (BRFSS) telephone survey.}
}
\value{
analysis List containing frequency and proportion data, plots,
    and statistical tests.
}
\description{
\code{analyzeRq3(brfss)}{Performs analyses for research question #3}
}
\details{
Function that performs the analysis for research question #3:  To what
degree does depression with co-occurring chronic illness affect productivity,
vis-a-vis that of diagnoses of depression or chronic illness separately?
}
\author{
John James, \email{jjames@datasciencestudio.org}
}
