% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRq4.R
\name{analyzeRq4}
\alias{analyzeRq4}
\title{analyzeRq4}
\usage{
analyzeRq4(brfss)
}
\arguments{
\item{brfss}{Data frame of preprocessed 2013 Behavioral Risk Factor
Surveillance System (BRFSS) telephone survey.}
}
\value{
analysis List containing frequency and proportion data, plots,
    and statistical tests.
}
\description{
\code{analyzeRq4(brfss)}{Performs analyses for research question #4}
}
\details{
Function that performs the analysis for research question #4:  Is there
a relationship between depression, chronic disease and the use of
health care services?
}
\author{
John James, \email{jjames@datasciencestudio.org}
}
