% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeUnivariate.R
\name{analyzeUnivariate}
\alias{analyzeUnivariate}
\title{analyzeUnivariate}
\usage{
analyzeUnivariate(brfss)
}
\arguments{
\item{brfss}{Data frame of preprocessed 2013 Behavioral Risk Factor Surveillance System (BRFSS) data}
}
\value{
univariate List of analyses for categorical and quantitative variables
}
\description{
\code{analyzeUnivariate}
}
\details{
This function performs the univariate analysis of categorical and quantitative
variables.
}
\author{
John James, \email{jjames@dataference.com}
}
