% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessData.R
\name{preProcessData}
\alias{preProcessData}
\title{preProcessData}
\usage{
preProcessData(brfss)
}
\arguments{
\item{brfss}{Data frame of 2013 Behavioral Risk Factor Surveillance System (BRFSS) data}
}
\value{
brfss Data frame containing variables required for downstream analysis
}
\description{
\code{preProcessData}
}
\details{
This function preprocesses the BRFSS data for downstream analysis.
}
\author{
John James, \email{jjames@dataference.com}
}
